# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GramSathi AI API
  description: |
    AI-Powered Multilingual Government Assistant API for rural citizens in India.
    Enables scheme discovery, application processing, document OCR, grievance filing,
    notifications, and analytics across 22 Indian languages.
  version: "1.0.0"
  contact:
    name: GramSathi Team
    email: dev@gramsathi.ai
  license:
    name: Proprietary
servers:
  - url: https://api.gramsathi.ai/v1
    description: Production server
  - url: https://staging-api.gramsathi.ai/v1
    description: Staging server
  - url: http://localhost:8001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtained from /auth/otp/verify
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        request_id:
          type: string
          format: uuid
          description: Unique request identifier for tracing
      required:
        - detail
...
    OTPSendRequest:
      type: object
      required:
        - mobile_number
      properties:
        mobile_number:
          type: string
          pattern: "^\\+?[1-9]\\d{9,14}$"
          example: "+919876543210"
          description: Mobile number with country code
    OTPSendResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "OTP sent successfully"
        expires_in:
          type: integer
          example: 300
        debug_otp:
          type: string
          nullable: true
          description: OTP value in development mode only
    OTPVerifyRequest:
      type: object
      required:
        - mobile_number
        - otp
      properties:
        mobile_number:
          type: string
          pattern: "^\\+?[1-9]\\d{9,14}$"
          example: "+919876543210"
        otp:
          type: string
          minLength: 4
          maxLength: 8
          example: "123456"
    OTPVerifyResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 900
        user:
          $ref: "#/components/schemas/UserOut"
    TokenRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    TokenRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
    LogoutRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    UserOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobile_number:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [citizen, admin, super_admin]
        is_verified:
          type: boolean
        preferred_language:
          type: string
          example: "hi"
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/otp/send:
    post:
      tags:
        - Authentication
      summary: Send OTP for login
      description: Generates and sends a one-time password to the citizen's mobile number via SMS/WhatsApp
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPSendRequest"
            example:
              mobile_number: "+919876543210"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OTPSendResponse"
              example:
                success: true
                message: "OTP sent successfully"
                expires_in: 300
                debug_otp: "123456"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      tags:
        - Authentication
      summary: Verify OTP and authenticate
      description: Verifies the OTP and returns JWT tokens for authenticated session
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPVerifyRequest"
            example:
              mobile_number: "+919876543210"
              otp: "123456"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OTPVerifyResponse"
              example:
                access_token: "eyJhbGciOiJSUzI1NiIs..."
                refresh_token: "eyJhbGciOiJSUzI1NiIs..."
                token_type: "bearer"
                expires_in: 900
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  mobile_number: "+919876543210"
                  name: null
                  role: "citizen"
                  is_verified: true
                  preferred_language: "hi"
                  created_at: "2026-07-04T10:00:00Z"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Uses a valid refresh token to issue a new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
            example:
              refresh_token: "eyJhbGciOiJSUzI1NiIs..."
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
              example:
                access_token: "eyJhbGciOiJSUzI1NiIs..."
                expires_in: 900
        "401":
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout and revoke session
      description: Revokes the refresh token, terminating the session
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              refresh_token: "eyJhbGciOiJSUzI1NiIs..."
      responses:
        "204":
          description: Logout successful, session revoked
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the authenticated user's profile information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                mobile_number: "+919876543210"
                name: "Ramesh Kumar"
                role: "citizen"
                is_verified: true
                preferred_language: "hi"
                created_at: "2026-07-04T10:00:00Z"

  # ===== Admin Auth =====
...